style: 볼트 환전 페이지 변경
