--- apps/admin/src/app/(main)/mentor/volt/exchange/page.tsx
@@ -1,5 +1,0 @@
-import ExchangeVolts from '@repo/admin/components/pages/main/mentor/volt/ExchangeVolts';
-
-export default function Page() {
-  return <ExchangeVolts />;
-}
