+++ apps/admin/src/components/pages/main/home/MentoringByMentor.tsx
@@ -49,9 +49,11 @@ function MentoringByMentor({
       setCheckedId([]);
     }
   };
+
   const handleOnClick = () => {
     router.push('/mentor/mentoring/add');
   };
+
   return (
     <div className="flex flex-col w-full">
       <CustomButton
