+++ packages/ui/src/components/ui/custom/CustomTableHeader.tsx
@@ -7,7 +7,7 @@ function CustomTableHeader({
   handleChange,
 }: {
   items: TableHeaderPropsType[];
-  handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  handleChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }) {
   return (
     <thead className="w-full bg-slate-200">
